Convolutive regression best model selection BIC

Convolutive regression best model selection BIC

BICselect(
  df,
  formula.resp,
  idx.pred,
  dist1,
  dist2,
  quiet = TRUE,
  formulas = "main",
  parallel = FALSE,
  interactions = FALSE,
  debug = FALSE,
  ...
)

Arguments

df
dataset

formula.resp
response formula

idx.pred
indices of predictors

dist1
chain of character to idnetify the distribution of variable1

dist2
chain of character to idnetify the distribution of variable2

quiet
logical. Displaying the progress.

formulas
"main" or "all". All includes the variances.

parallel
logical. Should the function run on parallel threads.

interactions
logical. Should the formulas include interactions.

debug
logical. Working option to debug. Dvp.

...
Additional arguments fed into convreg().

Examples

set.seed(123)
e=rnorm(n=50,mean=0,sd=0.05)
x = rnorm(50,0,0.5)
k= rnbinom(50,exp(x),0.5)
y= data.frame(obs= k + e +x, f1 = x, f2 = k, f3 = 2*rnorm(n=50,mean=0,sd=0.05))
form.r = ~obs
df = y
idx.pred = 2
dist1 = "Nbinom"
dist2 = "Gauss"
res= BICselect(df = df, formula.resp = form.r, idx.pred = idx.pred, dist1 = dist1, dist2 = dist2)

#> Number of combinations: 3

head(res,10)

#>   formula1 formula2        BIC
#> 1       ~1      ~f1   6.977019
#> 2      ~f1       ~1 197.063961
#> 3       ~1       ~1 199.857036